requirements.txt
streamlit==1.28.0
openai==1.3.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
gspread==5.12.0
pandas==2.1.4
requests==2.31.0